requirements.txt
torch
torchvision
ultralytics
faiss-cpu
#clip@
git+https://github.com/openai/CLIP.git
ftfy
regex
tqdm
transformers
opencv-python
Pillow
streamlit
fastapi
uvicorn
pytest
flake8
openai-whisper
moviepy